refactor: 초대 코드 공유 다이얼로그 통합 및 관련 로직 정리
- `ShareInviteCodeDialog.kt` 및 `ClipboardUtils.kt`를 삭제하고 `ShareInvitationCodeDialog.kt`로 기능을 통합합니다.
- `ShareInvitationCodeDialog` 내부에서 클립보드 복사 로직을 직접 처리하도록 변경하고 파라미터를 단순화합니다.
- `Context.sendSms` 확장 함수에 `ACTION_SENDTO`를 사용하고 예외 처리를 위한 `try-catch` 및 로그를 추가합니다.
- `JurorListScreen` 및 `NicknameScreen`에서 중복된 SMS 전송 로직을 제거하고 `Context.sendSms`를 사용하도록 수정합니다.
- `UserRepositoryImpl`의 초대 코드 조회 로직을 가독성 있게 개선합니다.
- `strings.xml`에서 중복된 초대 관련 문자열 리소스를 정리하고 형식을 통일합니다.
