+++ presentation/src/main/res/values/strings.xml
@@ -62,12 +62,7 @@
 
     <!-- Other -->
     <string name="back_pressed_message">뒤로가기를 한 번 더 누르면 종료됩니다.</string>
-    <string name="invite_dialog_title">내 초대코드를\n친구에게 공유해보세요</string>
-    <string name="invite_primary_click">문자로 공유하기</string>
-    <string name="invite_ghost_click">괜찮아요</string>
     <string name="dialog_cancel">취소</string>
-    <string name="invite_message">피클의 초대장이 도착했어요\n앱 설치 후, [심판 > 메이트 추가]에서 초대코드를 입력하세요.\n친구의 초대코드: [%1$s]\n\n서로의 소비를 판결하고, 함께 소비습관을 개선할 수 있어요.\n지금 바로 시작해보세요 ✨\n\n◆ pickle 다운로드\nhttps://play.google.com/store/apps/details?id=%2$s</string>
-
 
     <!-- Ledger -->
     <string name="ledger_empty_notice">가계부 내역을 작성해주세요</string>
@@ -180,9 +175,8 @@
     <string name="input_invitation_code_error_expired">만료된 초대코드입니다</string>
 
     <!-- ShareInvitationCode -->
-    <string name="share_invitation_code_dialog_title">친구에게 내 초대코드를\n공유해볼까요?</string>
+    <string name="share_invitation_code_dialog_title">내 초대코드를\n친구에게 공유해보세요</string>
     <string name="share_invitation_code_dialog_share_sms">문자로 공유하기</string>
     <string name="share_invitation_code_dialog_dismiss">괜찮아요</string>
-    <string name="share_invitation_sms_body">피클의 초대장이 도착했어요\n앱 설치 후, [심판 &gt; 메이트 추가]에서 초대코드를 입력하세요.\n친구의 초대코드: %1$s\n\n서로의 소비를 판결하고, 함께 소비습관을 개선할 수 있어요.\n지금 바로 시작해보세요 ✨\n\n◆ pickle 다운로드\nhttps://play.google.com/store/games?hl=ko</string>
-
+    <string name="share_invitation_sms_body">피클의 초대장이 도착했어요\n앱 설치 후, [심판 &gt; 메이트 추가]에서 초대코드를 입력하세요.\n친구의 초대코드: %1$s\n\n서로의 소비를 판결하고, 함께 소비습관을 개선할 수 있어요.\n지금 바로 시작해보세요 ✨\n\n◆ pickle 다운로드\nhttps://play.google.com/store/apps/details?id=%2$s</string>
 </resources>
